bugfix: 순환 참조 해결 및 Post/Comment 서비스 구조 개선
