+++ src/main/java/com/kakaotechcampus/team16be/comment/controller/CommentController.java
@@ -5,6 +5,7 @@ import com.kakaotechcampus.team16be.comment.dto.CommentIdResponse;
 import com.kakaotechcampus.team16be.comment.dto.CommentRequest;
 import com.kakaotechcampus.team16be.comment.dto.CommentResponse;
 import com.kakaotechcampus.team16be.comment.dto.CommentUpdateRequest;
+import com.kakaotechcampus.team16be.comment.service.CommentFacadeService;
 import com.kakaotechcampus.team16be.comment.service.CommentService;
 import com.kakaotechcampus.team16be.common.annotation.LoginUser;
 import com.kakaotechcampus.team16be.user.domain.User;
@@ -22,6 +23,7 @@ import java.util.List;
 public class CommentController {
 
     private final CommentService commentService;
+    private final CommentFacadeService commentFacadeService;
 
     @Operation(summary = "부모 댓글 작성", description = "부모 댓글을 작성합니다.(대댓글 X)")
     @PostMapping("/comments")
@@ -29,7 +31,7 @@ public class CommentController {
             @LoginUser User user,
             @RequestBody CommentRequest commentRequest
     ) {
-        Long commentId= commentService.createParentComment(user, commentRequest);
+        Long commentId= commentFacadeService.createParentComment(user, commentRequest);
         return ResponseEntity.ok(CommentIdResponse.from(commentId));
     }
 
@@ -39,7 +41,7 @@ public class CommentController {
             @LoginUser User user,
             @RequestBody CommentRequest commentRequest
     ) {
-        Long commentId = commentService.createChildComment(user, commentRequest);
+        Long commentId = commentFacadeService.createChildComment(user, commentRequest);
         return ResponseEntity.ok(CommentIdResponse.from(commentId));
     }
 
@@ -48,7 +50,7 @@ public class CommentController {
     public ResponseEntity<List<CommentResponse>> getAllComments(
             @PathVariable Long postId
     ) {
-        List<Comment> comments = commentService.getCommentsByPostId(postId);
+        List<Comment> comments = commentFacadeService.getCommentsByPostId(postId);
         return ResponseEntity.ok(CommentResponse.from(comments));
     }
