+++ src/main/java/com/kakaotechcampus/team16be/comment/service/CommentService.java
@@ -8,13 +8,10 @@ import com.kakaotechcampus.team16be.user.domain.User;
 import java.util.List;
 
 public interface CommentService {
-    Long createParentComment(User user,CommentRequest commentRequest);
-
-    Long createChildComment(User user,CommentRequest commentRequest);
-
-    List<Comment> getCommentsByPostId(Long postId);
 
     Long updateComment(User user, CommentUpdateRequest commentUpdateRequest);
 
     void deleteComment(User user, Long commentId);
+
+    Comment findById(Long id);
 }
