+++ src/main/java/com/kakaotechcampus/team16be/comment/service/CommentServiceImpl.java
@@ -20,39 +20,12 @@ import java.util.List;
 public class CommentServiceImpl implements CommentService {
 
     private final CommentRepository commentRepository;
-    private final PostService postService;
 
 
-    @Transactional
-    @Override
-    public Long createParentComment(User user, CommentRequest commentRequest) {
-        Post post = postService.findById(commentRequest.postId());
-
-        Comment parentComment = Comment.createComment(commentRequest.content(), post, user, null);
-        commentRepository.save(parentComment);
-        return parentComment.getId();
-    }
-
-    @Override
-    @Transactional
-    public Long createChildComment(User user,CommentRequest commentRequest) {
-        Post post = postService.findById(commentRequest.postId());
-
-        Comment comment = findById(commentRequest.parentId());
-
-        Comment childComment = Comment.createComment(commentRequest.content(), post, user, comment);
-
-        return commentRepository.save(childComment).getId();
-    }
 
     @Override
     @Transactional(readOnly = true)
-    public List<Comment> getCommentsByPostId(Long postId) {
-        Post post = postService.findById(postId);
-        return commentRepository.findAllByPost(post);
-    }
-
-    private Comment findById(Long id) {
+    public Comment findById(Long id) {
         return commentRepository.findById(id)
                 .orElseThrow(() -> new CommentException(CommentErrorCode.COMMENT_NOT_FOUND));
     }
