+++ src/main/java/com/kakaotechcampus/team16be/post/service/PostServiceImpl.java
@@ -1,6 +1,7 @@
 package com.kakaotechcampus.team16be.post.service;
 
 import com.kakaotechcampus.team16be.aws.service.S3UploadPresignedUrlService;
+import com.kakaotechcampus.team16be.comment.service.CommentFacadeService;
 import com.kakaotechcampus.team16be.comment.service.CommentService;
 import com.kakaotechcampus.team16be.group.domain.Group;
 import com.kakaotechcampus.team16be.group.service.GroupService;
@@ -29,9 +30,6 @@ public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
     private final GroupService groupService;
     private final GroupMemberService groupMemberService;
-    private final S3UploadPresignedUrlService s3UploadPresignedUrlService;
-    private final CommentService commentService;
-    private final PostLikeService postLikeService;
 
     @Override
     @Transactional
@@ -49,41 +47,6 @@ public class PostServiceImpl implements PostService {
         );
 
         return postRepository.save(post);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public GetPostResponse getPost(User user, Long groupId, Long postId) {
-        Group targetGroup = groupService.findGroupById(groupId);
-        Post post = postRepository.findByIdAndGroup(postId, targetGroup)
-                .orElseThrow(() -> new PostException(PostErrorCode.POST_NOT_FOUND));
-
-        Integer commentCount = commentService.getCommentsByPostId(postId).size();
-        PostLikeResponse postLikeResponse = postLikeService.getPostLikeInfo(user, postId);
-
-        List<String> fullURLs = post.getImageUrls().stream()
-                .map(s3UploadPresignedUrlService::getPublicUrl)
-                .toList();
-
-        return GetPostResponse.from(post, fullURLs, commentCount, postLikeResponse.isLiked());
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<GetPostResponse> getAllPosts(User user, Long groupId) {
-        Group targetGroup = groupService.findGroupById(groupId);
-        List<Post> posts = postRepository.findByGroup(targetGroup);
-
-        return posts.stream()
-                .map(post -> {
-                    List<String> fullURLs = post.getImageUrls().stream()
-                            .map(s3UploadPresignedUrlService::getPublicUrl)
-                            .toList();
-                    Integer commentCount = commentService.getCommentsByPostId(post.getId()).size();
-                    PostLikeResponse postLikeResponse = postLikeService.getPostLikeInfo(user, post.getId());
-                    return GetPostResponse.from(post, fullURLs, commentCount,postLikeResponse.isLiked());
-                })
-                .toList();
     }
 
     @Override
